fir_generator/requirements.txt
weasyprint==61.2
jinja2==3.1.3
pydantic==2.6.4
python-dateutil==2.9.0
supabase==2.4.1
python-dotenv==1.0.1
requests==2.31.0
groq==0.4.2
